Initial commit: Uploaded files
I did not include the cert.pem and key.pem, but you can generate your own if you want to run this on a local webserver. I used python's built in server for testing. The avatar will not show up on a locally ran index.html, it requires a webserver to run.